move ui useRender to interactions manager
